// src/modules/editor/core/schemas/short-answer.ts
import { z } from "zod";

export const shortAnswerPropsSchema = z.object({
  question: z.string().min(2).max(50),
  instructions: z.string().max(200),
  required: z.boolean().default(false),
  placeHolder: z.string().max(50),
});

export type ShortAnswerPropsSchemaType = z.infer<typeof shortAnswerPropsSchema>;